nodes are fine, roads need to be corrected in boolean is two way(not reading write), and subroads are wrong graph is creating design is wrong i think because of roads and subroads assign is wrong

+++ src/StreetMap.h
@@ -26,8 +26,8 @@ using namespace std;
 #define EARTH_RADIUS 6371.0
 
 class StreetMap {
-	map<int,Node> nodes;
-	map<int,Road> roads;
+	map<unsigned long int,Node> nodes;
+	map<unsigned long int,Road> roads;
 	Graph<int> graph;
 	//vector<Road> roads;
